+++ instrumentation/mysql2/lib/opentelemetry/instrumentation/mysql2.rb
@@ -5,6 +5,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 require 'opentelemetry'
+require 'opentelemetry-instrumentation-base'
 
 module OpenTelemetry
   module Instrumentation