+++ instrumentation/sidekiq/example/sidekiq.rb
@@ -5,6 +5,9 @@ require 'bundler/setup'
 
 Bundler.require
 
+# Export traces to console by default
+ENV['OTEL_TRACES_EXPORTER'] ||= 'console'
+
 OpenTelemetry::SDK.configure do |c|
   c.use 'OpenTelemetry::Instrumentation::Sidekiq'
 end