+++ common/CHANGELOG.md
@@ -1,5 +1,10 @@
 
 
+### v0.17.0 / 2021-04-22
+
+* ADDED: Add zipkin exporter 
+* ADDED: Processors validate exporters on init. 
+
 ### v0.16.0 / 2021-03-17
 
 * ADDED: Instrument lmdb gem