+++ instrumentation/lmdb/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release History: opentelemetry-instrumentation-lmdb
 
+### v0.17.0 / 2021-04-22
+
+* (No significant changes)
+
 ### v0.16.0 / 2021-03-17
 
 * Initial release.