+++ propagator/b3/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release History: opentelemetry-propagator-b3
 
+### v0.17.0 / 2021-04-22
+
+* BREAKING CHANGE: Replace TextMapInjector/TextMapExtractor pairs with a TextMapPropagator.
+
+  [Check the propagator documentation](https://open-telemetry.github.io/opentelemetry-ruby/) for the new usage.
+
+* FIXED: Refactor propagators to add #fields 
+
 ### v0.16.0 / 2021-03-17
 
 * DOCS: Replace Gitter with GitHub Discussions